+++ src/main/java/com/pp1/salve/api/loja/LojaController.java
@@ -1,13 +1,21 @@
 package com.pp1.salve.api.loja;
 
-import com.pp1.salve.model.loja.Loja;
-import com.pp1.salve.model.loja.LojaService;
+import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
+import com.pp1.salve.model.loja.Loja;
+import com.pp1.salve.model.loja.LojaService;
 
 @RestController
 @RequestMapping("/loja")
@@ -23,7 +31,7 @@ public class LojaController {
   }
 
   @GetMapping("/{id}")
-  public ResponseEntity<Loja> getById(@PathVariable Integer id) {
+  public ResponseEntity<Loja> getById(@PathVariable Long id) {
     return ResponseEntity.ok(service.findById(id));
   }
 
@@ -33,13 +41,13 @@ public class LojaController {
   }
 
   @PutMapping("/{id}")
-  public ResponseEntity<Loja> update(@PathVariable Integer id, @RequestBody Loja loja) {
+  public ResponseEntity<Loja> update(@PathVariable Long id, @RequestBody Loja loja) {
     loja.setId(id);
     return ResponseEntity.ok(service.save(loja));
   }
 
   @DeleteMapping("/{id}")
-  public ResponseEntity<Void> delete(@PathVariable Integer id) {
+  public ResponseEntity<Void> delete(@PathVariable Long id) {
     service.deleteById(id);
     return ResponseEntity.noContent().build();
   }
