+++ src/main/java/com/pp1/salve/api/pedido/PedidosController.java
@@ -1,13 +1,21 @@
 package com.pp1.salve.api.pedido;
 
-import com.pp1.salve.model.pedido.Pedido;
-import com.pp1.salve.model.pedido.PedidoService;
+import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
+import com.pp1.salve.model.pedido.Pedido;
+import com.pp1.salve.model.pedido.PedidoService;
 
 @RestController
 @RequestMapping("/pedidos")
@@ -23,7 +31,7 @@ public class PedidosController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Pedido> getById(@PathVariable Integer id) {
+    public ResponseEntity<Pedido> getById(@PathVariable Long id) {
         return ResponseEntity.ok(service.findById(id));
     }
 
@@ -33,13 +41,13 @@ public class PedidosController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Pedido> update(@PathVariable Integer id, @RequestBody Pedido pedido) {
+    public ResponseEntity<Pedido> update(@PathVariable Long id, @RequestBody Pedido pedido) {
         pedido.setId(id);
         return ResponseEntity.ok(service.save(pedido));
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Integer id) {
+    public ResponseEntity<Void> delete(@PathVariable Long id) {
         service.deleteById(id);
         return ResponseEntity.noContent().build();
     }
