+++ src/main/java/com/pp1/salve/model/loja/LojaRepository.java
@@ -2,6 +2,6 @@ package com.pp1.salve.model.loja;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LojaRepository extends JpaRepository<Loja, Integer> {
+public interface LojaRepository extends JpaRepository<Loja, Long> {
   
 }
